# packages/api/oas_v3.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "litefarm-api",
    "version": "3.2.1",
    "description": "Specification JSONs: [v2](/api-spec/v2), [v3](/api-spec/v3).\n\nLiteFarm API server"
  },
  "paths": {
...
    "/roles": {
      "get": {
        "summary": "/roles",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "role_id": {
                        "type": "number"
                      },
                      "role": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "role_id",
                      "role"
                    ]
                  }
                }
              }
            }
          },
          "304": {
            "description": ""
          }
        },
        "tags": [
          "roles"
        ]
      }
    },
...
    "/showed_spotlight": {
      "get": {
        "summary": "/showed_spotlight",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "map": {
                      "type": "boolean",
                      "example": false
                    },
                    "draw_area": {
                      "type": "boolean",
                      "example": false
                    },
                    "draw_line": {
                      "type": "boolean",
                      "example": false
                    },
                    "drop_point": {
                      "type": "boolean",
                      "example": false
                    },
                    "adjust_area": {
                      "type": "boolean",
                      "example": false
                    },
                    "adjust_line": {
                      "type": "boolean",
                      "example": false
                    },
                    "navigation": {
                      "type": "boolean",
                      "example": false
                    },
                    "notification": {
                      "type": "boolean",
                      "example": false
                    },
                    "introduce_map": {
                      "type": "boolean",
                      "example": true
                    },
                    "crop_catalog": {
                      "type": "boolean",
                      "example": false
                    },
                    "crop_variety_detail": {
                      "type": "boolean",
                      "example": false
                    },
                    "documents": {
                      "type": "boolean",
                      "example": false
                    },
                    "compliance_docs_and_certification": {
                      "type": "boolean",
                      "example": false
                    },
                    "transplant": {
                      "type": "boolean",
                      "example": false
                    },
                    "management_plan_creation": {
                      "type": "boolean",
                      "example": false
                    }
                  }
                }
              }
            }
          },
          "304": {
            "description": ""
          }
        },
        "tags": [
          "spotlight"
        ]
      },
      "patch": {
        "summary": "/showed_spotlight",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "OK"
                }
              }
            }
          }
        },
        "tags": [
          "spotlight"
        ]
      }
    },
...
    "/notification_user": {
      "get": {
        "summary": "/notification_user",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {}
                }
              }
            }
          },
          "304": {
            "description": ""
          }
        },
        "tags": [
          "notification",
          "user"
        ]
      },
      "patch": {
        "summary": "/notification_user",
        "responses": {},
        "tags": [
          "notification",
          "user"
        ]
      }
    },
...
  },
...
}